test(p2-shim): split browser suites
